server/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cloudinary==1.44.0
cryptography==45.0.3
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
idna==3.10
jwcrypto==1.5.6
Markdown==3.8
oauthlib==3.2.2
pillow==11.2.1
pycparser==2.22
python-dotenv==1.1.0
reportlab==4.4.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
